Rename widget-container to layout
- Rename widget-container.tsx → layout.tsx
- Rename widget-container.module.css → layout.module.css
- Rename .widgetContainer CSS class → .layout
- Update imports to use './layout'

File names now match the component name.
