+++ noodles-editor/src/layout.module.css
@@ -1,4 +1,4 @@
-.widgetContainer {
+.layout {
   display: grid;
   height: 100vh;
   grid-template-areas:
